+++ src/repl.rs
@@ -15,13 +15,114 @@
 
 //! Interactive interpreter for the EndBASIC language.
 
+use crate::ast::{ArgSep, Expr, VarType};
 use crate::console;
-use crate::exec::MachineBuilder;
+use crate::exec::{BuiltinCommand, Machine, MachineBuilder};
+use failure::Fallible;
 use rustyline::error::ReadlineError;
 use rustyline::Editor;
 use std::cell::RefCell;
+use std::collections::HashMap;
 use std::io::{self, Write};
 use std::rc::Rc;
+
+/// The `HELP` command.
+struct HelpCommand {
+    output: Rc<RefCell<dyn Write>>,
+}
+
+impl HelpCommand {
+    /// Prints a summary of all available commands.
+    fn summary(&self, builtins: &HashMap<&'static str, Rc<dyn BuiltinCommand>>) -> Fallible<()> {
+        let mut names = vec![];
+        let mut max_length = 0;
+        for name in builtins.keys() {
+            names.push(name);
+            if name.len() > max_length {
+                max_length = name.len();
+            }
+        }
+        names.sort();
+
+        let mut output = self.output.borrow_mut();
+        output.write_fmt(format_args!(
+            "\n    This is EndBASIC {}.\n\n",
+            env!("CARGO_PKG_VERSION")
+        ))?;
+        for name in names {
+            let filler = " ".repeat(max_length - name.len());
+            let builtin = builtins.get(name).unwrap();
+            let blurb = builtin.description().lines().next().unwrap();
+            output.write_fmt(format_args!(
+                "    {}{}    {}\n",
+                builtin.name(),
+                filler,
+                blurb
+            ))?;
+        }
+        output.write_all(
+            b"\n    Type HELP followed by a command name for details on that command.",
+        )?;
+        output.write_all(b"\n    Press CTRL+D to exit.\n\n")?;
+        Ok(())
+    }
+
+    /// Prints details about a single command.
+    fn describe(&self, builtin: &Rc<dyn BuiltinCommand>) -> Fallible<()> {
+        let mut output = self.output.borrow_mut();
+        output.write_all(b"\n")?;
+        if builtin.syntax().is_empty() {
+            output.write_fmt(format_args!("    {}\n", builtin.name()))?;
+        } else {
+            output.write_fmt(format_args!(
+                "    {} {}\n",
+                builtin.name(),
+                builtin.syntax()
+            ))?;
+        }
+        for line in builtin.description().lines() {
+            output.write_all(b"\n")?;
+            output.write_fmt(format_args!("    {}\n", line))?;
+        }
+        output.write_all(b"\n")?;
+        Ok(())
+    }
+}
+
+impl BuiltinCommand for HelpCommand {
+    fn name(&self) -> &'static str {
+        "HELP"
+    }
+
+    fn syntax(&self) -> &'static str {
+        "[commandname]"
+    }
+
+    fn description(&self) -> &'static str {
+        "Prints interactive help.
+Without arguments, shows a summary of all available commands.
+With a single argument, shows detailed information about the given command."
+    }
+
+    fn exec(&self, args: &[(Option<Expr>, ArgSep)], machine: &mut Machine) -> Fallible<()> {
+        let builtins = machine.get_builtins();
+        match args {
+            [] => {
+                self.summary(builtins)?;
+            }
+            [(Some(Expr::Symbol(vref)), ArgSep::End)] => {
+                ensure!(vref.ref_type() == VarType::Auto);
+                let name = vref.name().to_ascii_uppercase();
+                match &builtins.get(name.as_str()) {
+                    Some(builtin) => self.describe(builtin)?,
+                    None => bail!("Cannot describe unknown builtin {}", name),
+                }
+            }
+            _ => bail!("HELP takes zero or only one argument"),
+        }
+        Ok(())
+    }
+}
 
 /// Converts a `ReadLine` error into an `io::Error`.
 fn readline_error_to_io_error(e: ReadlineError) -> io::Error {
@@ -102,11 +203,15 @@ pub fn new_console() -> Rc<RefCell<dyn console::Console>> {
 pub fn run_repl_loop() -> io::Result<()> {
     let console = new_console();
     let mut machine = MachineBuilder::default()
+        .add_builtin(Rc::from(HelpCommand {
+            output: Rc::from(RefCell::from(io::stdout())),
+        }))
         .add_builtins(console::all_commands(console))
         .build();
 
     println!();
-    println!("    Welcome to EndBASIC {}", env!("CARGO_PKG_VERSION"));
+    println!("    Welcome to EndBASIC {}.", env!("CARGO_PKG_VERSION"));
+    println!("    Type HELP for interactive usage information.");
     println!();
 
     let mut rl = Editor::<()>::new();
@@ -134,4 +239,84 @@ pub fn run_repl_loop() -> io::Result<()> {
         }
     }
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    /// A command that does nothing.
+    struct DoNothingCommand {}
+
+    impl BuiltinCommand for DoNothingCommand {
+        fn name(&self) -> &'static str {
+            "DO_NOTHING"
+        }
+
+        fn syntax(&self) -> &'static str {
+            "this [would] <be|the> syntax \"specification\""
+        }
+
+        fn description(&self) -> &'static str {
+            "This is the blurb.
+First paragraph of the extended description.
+Second paragraph of the extended description."
+        }
+
+        fn exec(&self, _args: &[(Option<Expr>, ArgSep)], _machine: &mut Machine) -> Fallible<()> {
+            Ok(())
+        }
+    }
+
+    #[test]
+    fn test_help_summary() {
+        let output = Rc::from(RefCell::from(vec![]));
+        let mut machine = MachineBuilder::default()
+            .add_builtin(Rc::from(HelpCommand {
+                output: output.clone(),
+            }))
+            .add_builtin(Rc::from(DoNothingCommand {}))
+            .build();
+        machine.exec(&mut b"HELP".as_ref()).unwrap();
+
+        assert_eq!(
+            "
+    This is EndBASIC 0.1.0.
+
+    DO_NOTHING    This is the blurb.
+    HELP          Prints interactive help.
+
+    Type HELP followed by a command name for details on that command.
+    Press CTRL+D to exit.
+
+",
+            std::str::from_utf8(&output.borrow()).unwrap()
+        );
+    }
+
+    #[test]
+    fn test_help_describe() {
+        let output = Rc::from(RefCell::from(vec![]));
+        let mut machine = MachineBuilder::default()
+            .add_builtin(Rc::from(HelpCommand {
+                output: output.clone(),
+            }))
+            .add_builtin(Rc::from(DoNothingCommand {}))
+            .build();
+        machine.exec(&mut b"help Do_Nothing".as_ref()).unwrap();
+
+        assert_eq!(
+            "
+    DO_NOTHING this [would] <be|the> syntax \"specification\"
+
+    This is the blurb.
+
+    First paragraph of the extended description.
+
+    Second paragraph of the extended description.
+
+",
+            std::str::from_utf8(&output.borrow()).unwrap()
+        );
+    }
 }